# api/openapi.yaml
openapi: 3.0.0
info:
  title: LuminaBook API
  version: 1.0.0

paths:
  /ai/schedule:
    post:
      summary: Create meeting from natural language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
      responses:
        "200":
          description: Suggested time slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      type: object
                      properties:
                        startTime:
                          type: string
                        endTime:
                          type: string

  /events:
    post:
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startTime, endTime]
              properties:
                title:
                  type: string
                startTime:
                  type: string
                endTime:
                  type: string
      responses:
        "201":
          description: Event created

  /booking/{slug}:
    get:
      summary: Get available slots
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots

  /booking/{slug}/reserve:
    post:
      summary: Book a slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startTime, email]
              properties:
                startTime:
                  type: string
                email:
                  type: string
      responses:
        "201":
          description: Booking confirmed

  /user:
    get:
      operationId: user/get-users
      summary: Get a paginated list of users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        login:
                          type: string
                        firstName:
                          type: string
                        lastName:
                          type: string
                        email:
                          type: string
                        imageUrl:
                          type: string
                          nullable: true
                        activated:
                          type: boolean
                        langKey:
                          type: string
                        createdBy:
                          type: string
                        createdDate:
                          type: string
                          format: date-time
                        lastModifiedBy:
                          type: string
                        lastModifiedDate:
                          type: string
                          format: date-time
                        authorities:
                          type: array
                          items:
                            type: string
                  pageNumber:
                    type: integer
                  pageSize:
                    type: integer
                  pageCount:
                    type: integer
                  totalElements:
                    type: integer
        "401":
          description: Not authenticated
      tags:
        - User

  /user/{id}:
    get:
      operationId: user/get-user
      summary: Get a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  login:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  imageUrl:
                    type: string
                    nullable: true
                  activated:
                    type: boolean
                  langKey:
                    type: string
                  createdBy:
                    type: string
                  createdDate:
                    type: string
                    format: date-time
                  lastModifiedBy:
                    type: string
                  lastModifiedDate:
                    type: string
                    format: date-time
                  authorities:
                    type: array
                    items:
                      type: string
        "401":
          description: Not authenticated
        "404":
          description: User not found
      tags:
        - User

    delete:
      operationId: user/delete-user
      summary: Delete a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User deleted
        "401":
          description: Not authenticated
      tags:
        - User

  /account:
    get:
      operationId: account/get-account
      summary: Get the current user's account
      security:
        - bearerAuth: []
      parameters: []
      responses:
        "200":
          description: The authenticated user's account
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - login
                  - email
                  - activated
                properties:
                  id:
                    type: string
                  login:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  imageUrl:
                    type: string
                    nullable: true
                  activated:
                    type: boolean
                  langKey:
                    type: string
                  createdBy:
                    type: string
                  createdDate:
                    type: string
                    format: date-time
                  lastModifiedBy:
                    type: string
                  lastModifiedDate:
                    type: string
                    format: date-time
                  authorities:
                    type: array
                    items:
                      type: string
        "401":
          description: Not authenticated
        "500":
          description: User could not be resolved from authentication
      tags:
        - Account